remove MeshPool
